Use a different way to get the abbreviation
Because this doesn't assign snip.rv anymore, this solves the problem
encountered in the second test added in c89f7ec9 (cyclic dependency error)

The disadvantage is that now the whole line is seen as the abbreviation,
instead of only the matched regex of the snippet. But in practice, the
whole line will only consist of the Emmet abbreviation.